Fix cross-compiling the docs
docs.rs builds docs for all targets from a Linux host, so nothing that
uses a build.rs script that expects to run elsewhere will work.

Detect when we're cross-compiling and build stubs in place of
freebsd-libgeom-sys.  The stubs are good enough for freebsd-libgeom's
docs to build, but no better.

Also, test cross-compiling the docs in CI.